requirements.txt
requests
bs4
pydantic
python-dotenv